+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # GraphDB AWS Terraform Module Changelog
 
 ## 3.3.0
-
+* Added `graphdb_enable_audit_log_enabled` variable to enable GraphDB audit logging
 * Fixed perpetual Terraform drift in `aws_s3_bucket_server_side_encryption_configuration` for the backup module by explicitly setting `blocked_encryption_types = ["SSE-C"]` to match the value AWS returns
 
 ## 3.2.2